+++ src/App.tsx
@@ -1,18 +1,44 @@
 import './App.scss';
 import { NavBar } from './components/NavBar';
 import { Outlet } from 'react-router-dom';
-import { GlobalStateProvider } from './components/Store/PeopleStore';
+import {
+  DispatchContext,
+  GlobalStateProvider,
+  StateContext,
+} from './components/Store/PeopleStore';
+import { useContext, useEffect } from 'react';
+import { getPeople } from './api';
 
-export const App = () => (
-  <GlobalStateProvider>
-    <div data-cy="app">
-      <NavBar />
+export const App = () => {
+  const { people } = useContext(StateContext);
+  const dispatch = useContext(DispatchContext);
 
-      <main className="section">
-        <div className="container">
-          <Outlet />
-        </div>
-      </main>
-    </div>
-  </GlobalStateProvider>
-);
+  const loadPeople = async () => {
+    dispatch({ type: 'FETCH_START' });
+    try {
+      const result = await getPeople();
+
+      dispatch({ type: 'FETCH_SUCCESS', payload: result });
+    } catch {
+      dispatch({ type: 'FETCH_ERROR' });
+    }
+  };
+
+  useEffect(() => {
+    loadPeople();
+  }, [people]);
+
+  return (
+    <GlobalStateProvider>
+      <div data-cy="app">
+        <NavBar />
+
+        <main className="section">
+          <div className="container">
+            <Outlet />
+          </div>
+        </main>
+      </div>
+    </GlobalStateProvider>
+  );
+};
