+++ src/components/PeopleTable.tsx
@@ -1,26 +1,9 @@
-import { useContext, useEffect } from 'react';
+import { useContext } from 'react';
 import { PersonData } from './PersonData';
-import { DispatchContext, StateContext } from '../components/Store/PeopleStore';
-import { getPeople } from '../api';
+import { StateContext } from '../components/Store/PeopleStore';
 
 export const PeopleTable: React.FC = () => {
   const { people } = useContext(StateContext);
-  const dispatch = useContext(DispatchContext);
-
-  const loadPeople = async () => {
-    dispatch({ type: 'FETCH_START' });
-    try {
-      const result = await getPeople();
-
-      dispatch({ type: 'FETCH_SUCCESS', payload: result });
-    } catch {
-      dispatch({ type: 'FETCH_ERROR' });
-    }
-  };
-
-  useEffect(() => {
-    loadPeople();
-  }, [people]);
 
   return (
     <table
